@@ -33,83 +35,80 @@
   },
   "homepage": "https://github.com/ubigu/vuorovaikutusalusta#readme",
   "engines": {
-    "node": "^20.5.0"
+    "node": "^22.14.0"
   },
   "dependencies": {
-    "@date-io/date-fns": "^2.17.0",
+    "@date-io/date-fns": "^3.2.1",
     "@dnd-kit/core": "^6.3.1",
     "@dnd-kit/sortable": "^10.0.0",
     "@emotion/react": "^11.14.0",
     "@emotion/styled": "^11.14.0",
-    "@fontsource/montserrat": "^4.5.14",
-    "@fontsource/open-sans": "^4.5.14",
-    "@fontsource/roboto": "^4.5.8",
-    "@mui/icons-material": "^5.17.1",
-    "@mui/lab": "^5.0.0-alpha.176",
-    "@mui/material": "^5.17.1",
-    "@mui/styles": "^5.17.1",
-    "@mui/x-date-pickers": "^6.20.2",
-    "@turf/centroid": "^6.5.0",
+    "@fontsource/montserrat": "^5.2.6",
+    "@fontsource/open-sans": "^5.2.6",
+    "@fontsource/roboto": "^5.2.6",
+    "@mui/icons-material": "^7.1.2",
+    "@mui/lab": "^7.0.0-beta.14",
+    "@mui/material": "^7.1.2",
+    "@mui/x-date-pickers": "^8.5.3",
     "assert": "^2.1.0",
-    "date-fns": "^2.30.0",
+    "date-fns": "^4.1.0",
     "draft-js": "^0.11.7",
     "file-saver": "^2.0.5",
+    "globals": "^16.2.0",
     "jszip": "^3.10.1",
     "markdown-draft-js": "^2.4.0",
     "oskari-rpc": "^2.1.0",
     "parse-css-color": "^0.2.1",
     "process": "^0.11.10",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
     "react-draft-wysiwyg": "^1.15.0",
-    "react-dropzone": "^11.7.1",
-    "react-markdown": "^7.1.2",
+    "react-dropzone": "^14.3.8",
+    "react-markdown": "^10.1.0",
     "react-router-dom": "^5.3.4",
     "react-split-pane": "^0.1.92",
-    "recharts": "^2.15.4",
-    "rehype-external-links": "^1.0.1",
-    "remark-breaks": "^3.0.3"
+    "recharts": "^3.0.0",
+    "rehype-external-links": "^3.0.0",
+    "remark-breaks": "^4.0.0"
   },
   "devDependencies": {
     "@babel/preset-env": "^7.27.2",
     "@babel/preset-react": "^7.27.1",
+    "@testing-library/dom": "^10.4.0",
+    "@testing-library/react": "^16.3.0",
     "@types/draft-js": "^0.11.18",
-    "@types/enzyme": "^3.10.19",
     "@types/file-saver": "^2.0.7",
     "@types/geojson": "^7946.0.16",
-    "@types/jest": "^28.1.8",
+    "@types/jest": "^30.0.0",
     "@types/markdown-draft-js": "^2.2.7",
-    "@types/node": "^20.19.1",
-    "@types/react": "^17.0.87",
-    "@types/react-dom": "^17.0.26",
+    "@types/node": "^24.0.3",
+    "@types/react": "^19.1.8",
+    "@types/react-dom": "^19.1.6",
     "@types/react-draft-wysiwyg": "^1.13.8",
     "@types/react-router-dom": "^5.3.3",
-    "@typescript-eslint/eslint-plugin": "^4.33.0",
-    "@typescript-eslint/parser": "^4.33.0",
+    "@typescript-eslint/eslint-plugin": "^8.35.0",
+    "@typescript-eslint/parser": "^8.35.0",
     "@vitejs/plugin-react": "^4.6.0",
-    "@wojtekmaj/enzyme-adapter-react-17": "^0.8.0",
     "babel-eslint": "^10.1.0",
-    "dotenv": "^10.0.0",
-    "enzyme": "^3.11.0",
-    "eslint": "^7.32.0",
-    "eslint-config-standard": "^16.0.3",
+    "dotenv": "^16.5.0",
+    "eslint": "^9.29.0",
     "eslint-plugin-import": "^2.32.0",
     "eslint-plugin-node": "^11.1.0",
-    "eslint-plugin-promise": "^5.2.0",
+    "eslint-plugin-promise": "^7.2.1",
     "eslint-plugin-react": "^7.37.5",
     "eslint-plugin-standard": "^5.0.0",
     "glob-parent": "^6.0.2",
     "inline-source-map": "^0.6.3",
-    "jest": "^28.1.3",
+    "jest": "^30.0.2",
     "jest-canvas-mock": "^2.5.2",
-    "jest-environment-jsdom": "^28.1.3",
+    "jest-environment-jsdom": "^30.0.2",
     "jest-fetch-mock": "^3.0.3",
-    "prettier-plugin-organize-imports": "^2.3.4",
-    "ts-jest": "^28.0.8",
-    "typescript": "^4.9.5",
+    "prettier-plugin-organize-imports": "^4.1.0",
+    "ts-jest": "^29.4.0",
+    "typescript": "^5.8.3",
     "url-parse": "^1.5.10",
-    "vite": "^4.5.14",
-    "vite-tsconfig-paths": "^4.3.2"
+    "vite": "^6.3.5",
+    "vite-tsconfig-paths": "^5.1.4"
   },
   "overrides": {
     "react-split-pane": {
