@@ -1,11 +1,10 @@
-import { Button } from '@mui/material';
-import { makeStyles } from '@mui/styles';
+import { Box, Button } from '@mui/material';
 import { useTranslations } from '@src/stores/TranslationContext';
 import React, { useState } from 'react';
 import SaveAsUnfinishedDialog from './SaveAsUnfinishedDialog';
 import { SurveyPage } from '@interfaces/survey';
 
-const useStyles = makeStyles({
+const styles = {
   root: {
     display: 'flex',
     flexDirection: 'row',
@@ -13,7 +12,7 @@ const useStyles = makeStyles({
       marginRight: '0.5rem',
     },
   },
-});
+};
 
 interface Props {
   activeStep: number;
@@ -34,11 +33,10 @@ interface Props {
 export default function StepperControls(props: Props) {
   const [saveDialogOpen, setSaveDialogOpen] = useState(false);
   const { tr } = useTranslations();
-  const classes = useStyles();
 
   return (
     <>
-      <div className={classes.root}>
+      <Box sx={styles.root}>
         {props.activeStep > 0 && (
           <Button
             disabled={props.disabled}
@@ -80,7 +78,7 @@ export default function StepperControls(props: Props) {
             {tr.SurveyStepper.saveAsUnfinished}
           </Button>
         )}
-      </div>
+      </Box>
       {props.allowSavingUnfinished && (
         <SaveAsUnfinishedDialog
           open={saveDialogOpen}
