@@ -1,9 +1,8 @@
 import { Box, Link, Typography } from '@mui/material';
-import { makeStyles } from '@mui/styles';
 import { useTranslations } from '@src/stores/TranslationContext';
 import React from 'react';
 
-const useStyles = makeStyles({
+const styles = {
   infoContainer: {
     margin: '0rem 1rem 0rem 1rem',
     display: 'flex',
@@ -13,16 +12,15 @@ const useStyles = makeStyles({
     fontSize: '1.25rem',
     paddingBottom: '0.5rem',
   },
-});
+};
 
 export function UnavailableSurvey() {
   const { tr } = useTranslations();
-  const classes = useStyles();
 
   return (
     <Box
-      className={classes.infoContainer}
       sx={{
+        ...styles.infoContainer,
         display: 'flex',
         height: '75vh',
         alignItems: 'center',
@@ -35,14 +33,14 @@ export function UnavailableSurvey() {
         alt={tr.IconAltTexts.logoAltText}
       />
       <br />
-      <Typography variant="body1" classes={{ body1: classes.infoText }}>
+      <Typography variant="body1" sx={styles.infoText}>
         {tr.UnavailableSurvey.surveyNotPublished}
       </Typography>
-      <Typography variant="body1" classes={{ body1: classes.infoText }}>
+      <Typography variant="body1" sx={styles.infoText}>
         {tr.UnavailableSurvey.treTopicalInfo}
       </Typography>
       <br />
-      <Link href="https://tampere.fi" className={classes.infoText}>
+      <Link href="https://tampere.fi" sx={styles.infoText}>
         {tr.UnavailableSurvey.tre}
       </Link>
     </Box>
