@@ -1,24 +1,21 @@
-import { makeStyles } from '@mui/styles';
+import { Box } from '@mui/material';
 import React from 'react';
 
 interface Props {
   color: string;
 }
 
-const useStyles = makeStyles({
+const styles = {
   colorIndicator: {
     borderRadius: '50%',
     width: '1rem',
     height: '1rem',
     marginLeft: '0.5rem',
   },
-});
+};
 
 export default function ColorIndicator({ color }: Props) {
-  const classes = useStyles();
   return (
-    color && (
-      <div className={classes.colorIndicator} style={{ background: color }} />
-    )
+    color && <Box sx={styles.colorIndicator} style={{ background: color }} />
   );
 }
