@@ -273,7 +273,9 @@ export function EditSurveyPageConditions() {
               value={
                 conditionList.length > 0
                   ? conditionList[0][0]
-                  : String(previousQuestions?.[0].id) ?? ''
+                  : previousQuestions?.[0].id
+                  ? String(previousQuestions?.[0].id)
+                  : ''
               }
               onChange={(event) => handleQuestionSelect(event)}
             >
